002 - Review OK

+++ thisfinans/src/main/resources/db/migration/V2__add-fk-account-transactions.sql
@@ -1,3 +1,3 @@
 ALTER TABLE TRANSACTIONS
-    ADD account_fk INT NOT NULL,
+    ADD account_fk BIGINT NOT NULL,
     ADD CONSTRAINT fk_transactions_account_id FOREIGN KEY (account_fk) REFERENCES ACCOUNTS (account_id);
