+++ thisfinans/src/main/resources/db/migration/V4__add-fk-account-transactionsEntries.sql
@@ -1,3 +1,3 @@
 ALTER TABLE TRANSACTION_ENTRIES
-    ADD account_fk INT NOT NULL,
+    ADD account_fk BIGINT NOT NULL,
     ADD CONSTRAINT fk_transaction_entries_account_id FOREIGN KEY (account_fk) REFERENCES ACCOUNTS (account_id);
